pachyderm-seldon/experiment/requirements.txt
# Requirements for the IDE

determined==0.18.0
torch==1.10.2
Pillow==9.0.1
numpy==1.22.2
torchvision==0.11.3
scikit-image==0.19.2
python_pachyderm==7.1.1
